+++ monai/networks/blocks/dynunet_block.py
@@ -21,7 +21,7 @@ from monai.networks.layers.factories import Act, Norm, split_args
 
 class UnetResBlock(nn.Module):
     """
-    A skip-connection based module that can be used for NNUNet, based on:
+    A skip-connection based module that can be used for DynUNet, based on:
     `Automated Design of Deep Learning Methods for Biomedical Image Segmentation <https://arxiv.org/abs/1904.08128>`_.
     `nnU-Net: Self-adapting Framework for U-Net-Based Medical Image Segmentation <https://arxiv.org/abs/1809.10486>`_.
 
@@ -96,7 +96,7 @@ class UnetResBlock(nn.Module):
 
 class UnetBasicBlock(nn.Module):
     """
-    A CNN module module that can be used for NNUNet, based on:
+    A CNN module module that can be used for DynUNet, based on:
     `Automated Design of Deep Learning Methods for Biomedical Image Segmentation <https://arxiv.org/abs/1904.08128>`_.
     `nnU-Net: Self-adapting Framework for U-Net-Based Medical Image Segmentation <https://arxiv.org/abs/1809.10486>`_.
 
@@ -153,7 +153,7 @@ class UnetBasicBlock(nn.Module):
 
 class UnetUpBlock(nn.Module):
     """
-    An upsampling module that can be used for NNUNet, based on:
+    An upsampling module that can be used for DynUNet, based on:
     `Automated Design of Deep Learning Methods for Biomedical Image Segmentation <https://arxiv.org/abs/1904.08128>`_.
     `nnU-Net: Self-adapting Framework for U-Net-Based Medical Image Segmentation <https://arxiv.org/abs/1809.10486>`_.
