+++ monai/networks/nets/__init__.py
@@ -12,9 +12,9 @@
 from .ahnet import AHNet
 from .classifier import *
 from .densenet import DenseNet, densenet121, densenet169, densenet201, densenet264
+from .dynunet import DynUNet
 from .generator import Generator
 from .highresnet import HighResBlock, HighResNet
-from .nnunet import NNUnet
 from .regressor import Regressor
 from .segresnet import SegResNet, SegResNetVAE
 from .senet import SENet, se_resnet50, se_resnet101, se_resnet152, se_resnext50_32x4d, se_resnext101_32x4d, senet154
